==> 1_Generation_Train_Test_HoldOut_Set.ipynb <==
import numpy as np
import pandas as pd
import random

from surprise import SVD
from surprise import SVDpp
from surprise import NMF
from surprise import KNNBasic
from surprise import KNNWithZScore
from surprise import Dataset
from surprise import Reader
from surprise.model_selection import cross_validate

from surprise.model_selection import GridSearchCV

# For repeatability, set the seed to a default value
my_seed = 0
random.seed(my_seed)
np.random.seed(my_seed)

# Reading all data sets -- This contain all the Train data set that will be used
# for analysis

All_Data = pd.read_csv('./Files_Folder/All_Data_Set.csv')
All_Data.head(2)

# Dividing data into train - test - holdout
hold_data = All_Data.sample(frac = 0.05, random_state = 0)
train_test_set = All_Data.drop(hold_data.index)
test_data = train_test_set.sample(hold_data.shape[0], random_state = 0)
train_data = train_test_set.drop(test_data.index)

print ('Shape of All Data = ', All_Data.shape)
print ('Shape of Train Data = ', train_data.shape)
print ('Shape of Test Data = ', test_data.shape)
print ('Shape of Hold-Out Data = ', hold_data.shape)

All_Data.head(2)

train_data.head()

test_data.head()

hold_data.head()

# Checking out how the data set have been divided into portions
test_data_sub = test_data.drop(['Purchase'], axis=1)
hold_data_sub = hold_data.drop(['Purchase'], axis=1)

print ('Shape of All Data = ', All_Data.shape)
print ('Shape of Train Data = ', train_data.shape)
print ('Shape of Test Data (No Purchase column) = ', test_data_sub.shape)
print ('Shape of Hold-Out Data (No Purchase column) = ', hold_data_sub.shape)

test_data_sub.head()

hold_data_sub.head()

# Saving all the data sets to different files ...
train_data.to_csv ('./Files_Folder/train_data.csv', index=False) # For model development.
test_data_sub.to_csv('./Files_Folder/test_data_sub.csv', index=False)
hold_data_sub.to_csv ('./Files_Folder/hold_data_sub.csv', index=False)
test_data.to_csv('./Files_Folder/test_data.csv', index=False) # For model optimization
hold_data.to_csv ('./Files_Folder/hold_data.csv', index=False) # Final answer for the hold out data will be!

# One needs to determine the Product_IDs that are present in the test data set but not present in the train data set
