==> 2_New_Prod_IDs.ipynb <==
import numpy as np
import pandas as pd
import random

from surprise import SVD
from surprise import SVDpp
from surprise import NMF
from surprise import KNNBasic
from surprise import KNNWithZScore
from surprise import Dataset
from surprise import Reader
from surprise.model_selection import cross_validate

from surprise.model_selection import GridSearchCV

from sklearn import preprocessing

# Reading the data sets into data frames
train_data = pd.read_csv('./Files_Folder/train_data.csv')
test_data = pd.read_csv('./Files_Folder/test_data.csv')
test_data_sub = pd.read_csv('./Files_Folder/test_data_sub.csv')
hold_data = pd.read_csv('./Files_Folder/hold_data.csv')
hold_data_sub = pd.read_csv('./Files_Folder/hold_data_sub.csv')

print ("train shape = ", train_data.shape)
print ("test shape = ", test_data.shape)
print ("test shape (No Purchase column) = ", test_data_sub.shape)
print ("holdout shape = ", hold_data.shape)
print ("holdout shape (No Purchase column) = ", hold_data_sub.shape)

train_data.head(5)

train_data.describe()

# Checking out features that have missing values for the train data set
train_data.isnull().sum()

# Checking out the statistical description of numeric features for the test data set
test_data_sub.describe()

# Checking out the features that have missing values in the test data set
test_data_sub.isnull().sum()

hold_data_sub.describe()

# Checking out the features that have missing values in the hold out data set
hold_data_sub.isnull().sum()

# Filling missing values with place holders
train_data.Product_Category_2.fillna('BB', inplace = True)
train_data.Product_Category_3.fillna('CC', inplace = True)
test_data_sub.Product_Category_2.fillna('BB', inplace = True)
test_data_sub.Product_Category_3.fillna('CC', inplace = True)
hold_data_sub.Product_Category_2.fillna('BB', inplace = True)
hold_data_sub.Product_Category_3.fillna('CC', inplace = True)

train_data.isnull().sum()

test_data_sub.isnull().sum()

hold_data_sub.isnull().sum()

# Checking the number of unique Product_IDs in the test, train & hold out data sets
train_Prod_IDs = list(set(train_data.Product_ID))
test_Prod_IDs = list(set(test_data_sub.Product_ID))
hold_Prod_IDs = list(set(hold_data_sub.Product_ID))
print (len(train_Prod_IDs))
print (len(test_Prod_IDs))
print (len(hold_Prod_IDs))

nid = test_Prod_IDs
print ('length of test_IDs = ', len (nid))

nid.extend(hold_Prod_IDs)
print ('length of test_IDs & holdout_IDs = ', len(nid))

nid = set(nid)
print ('length of unique test_IDs & holdout_IDs = ', len(nid) )

New_Prod_IDs = []
for i in nid:
    if i not in train_Prod_IDs:
        New_Prod_IDs.append(i)

print (len(New_Prod_IDs))
New_Prod_IDs

# Putting test and hold in a single dataframe
print (test_data_sub.shape)
print (hold_data_sub.shape)
frames = [test_data_sub, hold_data_sub]
test_hold_data_sub = pd.concat(frames)
print (test_hold_data_sub.shape)

# Getting the summary of the products in train data set that has the same product categories as 
# the new products in test data sets.
# New products in test data sets are products that were not in the train data sets but are found in
# test data sets

agg_mean = []
agg_median = []
agg_std = []
agg_count = []
agg_min = []
agg_max = []
agg_25 = []
agg_75 = []

for i in New_Prod_IDs:
    
    a = test_hold_data_sub[test_hold_data_sub.Product_ID == i].iloc[0]
    agg_mean.append(train_data[(train_data.Product_Category_1 == a.Product_Category_1) & (train_data.Product_Category_2 == a.Product_Category_2) & (train_data.Product_Category_3 == a.Product_Category_3)].describe().Purchase['mean'])
    agg_median.append(train_data[(train_data.Product_Category_1 == a.Product_Category_1) & (train_data.Product_Category_2 == a.Product_Category_2) & (train_data.Product_Category_3 == a.Product_Category_3)].describe().Purchase['50%'])
    agg_std.append(train_data[(train_data.Product_Category_1 == a.Product_Category_1) & (train_data.Product_Category_2 == a.Product_Category_2) & (train_data.Product_Category_3 == a.Product_Category_3)].describe().Purchase['std'])
    agg_count.append(train_data[(train_data.Product_Category_1 == a.Product_Category_1) & (train_data.Product_Category_2 == a.Product_Category_2) & (train_data.Product_Category_3 == a.Product_Category_3)].describe().Purchase['count'])
    agg_min.append(train_data[(train_data.Product_Category_1 == a.Product_Category_1) & (train_data.Product_Category_2 == a.Product_Category_2) & (train_data.Product_Category_3 == a.Product_Category_3)].describe().Purchase['min'])
    agg_max.append(train_data[(train_data.Product_Category_1 == a.Product_Category_1) & (train_data.Product_Category_2 == a.Product_Category_2) & (train_data.Product_Category_3 == a.Product_Category_3)].describe().Purchase['max'])
    agg_25.append(train_data[(train_data.Product_Category_1 == a.Product_Category_1) & (train_data.Product_Category_2 == a.Product_Category_2) & (train_data.Product_Category_3 == a.Product_Category_3)].describe().Purchase['25%'])
    agg_75.append(train_data[(train_data.Product_Category_1 == a.Product_Category_1) & (train_data.Product_Category_2 == a.Product_Category_2) & (train_data.Product_Category_3 == a.Product_Category_3)].describe().Purchase['75%'])

  

agg_dic = {'Product_ID':New_Prod_IDs, 'mean': agg_mean, '50%': agg_median, 'std': agg_std, 'count':agg_count, 'min':agg_min, 'max':agg_max, '25%': agg_25, '75%': agg_75}    

# Storing the details of similar products in a csv file
# This can be used as a look up table during the analysis.
# The Product_ID are products that are in the test data set but not in the train data set
# The other parameters like mean, 50% etc are the summary of purchases of similar products 
# found in the train data set
New_Product = pd.DataFrame (agg_dic)
New_Product.head()

New_Product.to_csv('./Files_Folder/New_Product_Details.csv', index = False)

train_data.head(2)

# Creating a pivot Table from User & Product IDs
train_User_Prod = pd.pivot_table(train_data, values = 'Purchase', index = 'User_ID', columns = 'Product_ID')
train_User_Prod.head(2)

train_User_Prod.to_csv('./Files_Folder/tr_User_Prod_Original_Pivot.csv')

# The statistical description of each of the Product_ID
# The minimum and maximum for each Product_ID can be used for reversing the normalization step after prediction
tr_Us_Pr_statistics = train_User_Prod.describe()
tr_Us_Pr_statistics

# Saving the Train_User_Product_Statistics file as a csv
tr_Us_Pr_statistics.to_csv('./Files_Folder/train_User_Prod_stat.csv')

# Since all the counts add up to (90% of 500068), it imples that there is no product that was purchased 
# by two User_IDs.  
tr_Us_Pr_statistics.loc['count', :].sum()

# Normalizing the Purchases
minmax = preprocessing.MinMaxScaler(feature_range=(1,5))
train_User_Prod_Scaled = minmax.fit_transform(train_User_Prod)

train_User_Prod_Scaled

x_row = list(train_User_Prod.index)
x_col = list(train_User_Prod.columns)
train_User_Prod_Scaled_df = pd.DataFrame(train_User_Prod_Scaled, index = x_row, columns = x_col)

train_User_Prod_Scaled_df.head(3)

train_User_Prod_Scaled_df.describe()

train_User_Prod_Scaled_df.head()

train_User_Prod.head()

train_User_Prod_Scaled_df.to_csv('./Files_Folder/tr_User_Prod_Normalized_Pivot.csv')

# Preparing Data in the form acceptable by Surprise Package

train_data.head()

User_ID_list = list(train_data.User_ID)
Prod_ID_list = list(train_data.Product_ID)

# Creating a list of normalized purchases in the same order as listed in the train data set
Purc_list = []
for i in range(len(User_ID_list)):
    Purc_list.append(train_User_Prod_Scaled_df.loc[User_ID_list[i], Prod_ID_list[i]])

User_Prod_NormPurc = {'User_ID': User_ID_list, 'Product_ID': Prod_ID_list, 'Normalized_Purchase': Purc_list}

User_Prod_NormPurc_df = pd.DataFrame(User_Prod_NormPurc)

User_Prod_NormPurc_df.head()

User_Prod_NormPurc_df.describe()

User_Prod_NormPurc_df.to_csv('./Files_Folder/User_Prod_NormPurc.csv', index=False)
