==> 6_Error_Analysis.ipynb <==
import numpy as np
import pandas as pd
import random

from surprise import SVD
from surprise import SVDpp
from surprise import NMF
from surprise import KNNBasic
from surprise import KNNWithZScore
from surprise import Dataset
from surprise import Reader
from surprise.model_selection import cross_validate

from surprise.model_selection import GridSearchCV

my_seed = 0
random.seed(my_seed)
np.random.seed(my_seed)

# train and test data sets
train_data = pd.read_csv('./Files_Folder/train_data.csv')

test_data = pd.read_csv('./Files_Folder/test_data.csv')

print (train_data.shape)
print (test_data.shape)

# Reading all the predicted values
train_data_pred = pd.read_csv('./Files_Folder/Predicted_train_Result_SVD.csv')
test_data_mean_pred = pd.read_csv('./Files_Folder/Predicted_test_SVD_mean.csv')
test_data_median_pred = pd.read_csv('./Files_Folder/Predicted_test_SVD_median.csv')

print (train_data_pred.shape)
print (test_data_mean_pred.shape)
print (test_data_median_pred.shape)

train_data_pred.dtypes

train_data_pred.columns = ['User_ID', 'Product_ID', 'Predicted_Purchase']
test_data_mean_pred.columns = ['User_ID', 'Product_ID', 'Predicted_Purchase']
test_data_median_pred.columns = ['User_ID', 'Product_ID', 'Predicted_Purchase']

train_data_pred['Actual_Purchase'] = train_data['Purchase']
test_data_mean_pred['Actual_Purchase'] = test_data['Purchase']
test_data_median_pred['Actual_Purchase'] = test_data['Purchase']

train_data_pred['Error'] = train_data_pred['Predicted_Purchase'] - train_data_pred['Actual_Purchase']
test_data_mean_pred['Error'] = test_data_mean_pred['Predicted_Purchase'] - test_data_mean_pred['Actual_Purchase']
test_data_median_pred['Error'] = test_data_median_pred['Predicted_Purchase'] - test_data_median_pred['Actual_Purchase']

train_data_pred['Error_sq'] = train_data_pred['Error']*train_data_pred['Error']
test_data_mean_pred['Error_sq'] = test_data_mean_pred['Error']*test_data_mean_pred['Error']
test_data_median_pred['Error_sq'] = test_data_median_pred['Error']*test_data_median_pred['Error']

rmse_train_data_pred = np.sqrt(train_data_pred.Error_sq.sum()/train_data_pred.shape[0])
rmse_test_data_mean_pred = np.sqrt(test_data_mean_pred.Error_sq.sum()/test_data_mean_pred.shape[0])
rmse_test_data_median_pred = np.sqrt(test_data_median_pred.Error_sq.sum()/test_data_median_pred.shape[0])

print (rmse_train_data_pred)
print (rmse_test_data_mean_pred)
print (rmse_test_data_median_pred)

test_data_mean_pred.shape

test_data_mean_pred_2 = test_data_mean_pred.drop(test_data_newid.index)
test_data_mean_pred_2.shape

np.sqrt(test_data_mean_pred_2.Error_sq.sum()/test_data_mean_pred_2.shape[0])

train_data.describe()
