==> 7_Varying_reg_param.ipynb <==
import numpy as np
import pandas as pd
import random

from surprise import SVD
from surprise import SVDpp
from surprise import NMF
from surprise import KNNBasic
from surprise import KNNWithZScore
from surprise import Dataset
from surprise import Reader
from surprise.model_selection import cross_validate

from surprise.model_selection import GridSearchCV

my_seed = 0
random.seed(my_seed)
np.random.seed(my_seed)

# Train data set with Normalized Purchase
User_Prod_NP = pd.read_csv('./Files_Folder/User_Prod_NormPurc.csv')

# Statistics of each Product_ID in train data sets
train_Prod_sta = pd.read_csv('./Files_Folder/train_User_Prod_stat.csv') 

# Statistics of new Product_IDs; those are Prodcut_IDs in test data set but not in train data set
New_Prod_LUT = pd.read_csv('./Files_Folder/New_Product_Details.csv')

# Test Data Set
test_data_sub = pd.read_csv('./Files_Folder/test_data_sub.csv')

# Hold Out Data Set
hold_data_sub = pd.read_csv('./Files_Folder/hold_data_sub.csv')

# Train Data Set
train_data = pd.read_csv('./Files_Folder/train_data.csv')

test_data = pd.read_csv('./Files_Folder/test_data.csv')

new_prod = list(New_Prod_LUT.Product_ID)

cv_mean = []
cv_std = []
rmse_train = []
rmse_test = []

for j in range(20):
    
    j1 = 0.01*(j+1)
    algo_SVD = SVD(n_epochs=130, lr_all = 0.005, reg_all = j1)

    # A reader is still needed but only the rating_scale param is requiered.
    reader = Reader(rating_scale=(1, 5))

    # The columns must correspond to user id, item id and ratings (in that order).
    data = Dataset.load_from_df(User_Prod_NP[['User_ID', 'Product_ID', 'Normalized_Purchase']], reader)

    # We can now use this dataset as we please, e.g. calling cross_validate
    crva = cross_validate(algo_SVD, data, cv=3, verbose = True)

    cv_mean.append(crva['test_rmse'].mean())

    cv_std.append(crva['test_rmse'].std())
    
    # Using our model to predict the prices of the train data set
    # This is used to determine the training error
    train_user_list = []
    train_prod_list = []
    train_price = []
    for i in range(train_data.shape[0]):
        p1 = algo_SVD.predict(train_data.User_ID[i],train_data.Product_ID[i]).est
        rmin = train_Prod_sta.loc[3, train_data.Product_ID[i]]
        rmax = train_Prod_sta.loc[7, train_data.Product_ID[i]]
        dmin = 1
        dmax = 5
        p = (p1 - dmin)*(rmax - rmin)/(dmax - dmin) + rmin
        train_price.append(p)
        train_user_list.append(train_data.User_ID[i])
        train_prod_list.append(train_data.Product_ID[i])
    final_train_tally = {'User_ID':train_user_list, 'Product_ID':train_prod_list, 'Purchase': train_price}
    final_train_tally_df = pd.DataFrame(final_train_tally)
    
    
    test_user_list = []
    test_prod_list = []
    price = []
    for i in range(test_data_sub.shape[0]):
        if test_data_sub.Product_ID[i] in new_prod:
            p = New_Prod_LUT[New_Prod_LUT.Product_ID == test_data_sub.Product_ID[i]]['mean'].iloc[0]      
        else:
            p1 = algo_SVD.predict(test_data_sub.User_ID[i],test_data_sub.Product_ID[i]).est
            rmin = train_Prod_sta.loc[3, test_data_sub.Product_ID[i]]
            rmax = train_Prod_sta.loc[7, test_data_sub.Product_ID[i]]
            dmin = 1
            dmax = 5
            p = (p1 - dmin)*(rmax - rmin)/(dmax - dmin) + rmin
        price.append(p)
        test_user_list.append(test_data_sub.User_ID[i])
        test_prod_list.append(test_data_sub.Product_ID[i])
    final_test_tally = {'User_ID':test_user_list, 'Product_ID':test_prod_list, 'Purchase': price}
    final_test_tally_df = pd.DataFrame(final_test_tally)
    
    final_train_tally_df ['Actual_Purchase'] = train_data['Purchase']
    final_train_tally_df ['Error'] = final_train_tally_df ['Purchase'] - final_train_tally_df ['Actual_Purchase']
    final_train_tally_df ['Error_sq'] = final_train_tally_df ['Error']*final_train_tally_df ['Error']
    rmse_train.append(np.sqrt(final_train_tally_df ['Error_sq'].sum()/final_train_tally_df.shape[0]))    
    file_train = './Files_Folder/rmse/reg/Predicted_train_Result_SVD_mean_' + str(j+1) + '.csv'
    final_train_tally_df.to_csv(file_train, index=False) 
    
    final_test_tally_df ['Actual_Purchase'] = test_data['Purchase']
    final_test_tally_df ['Error'] = final_test_tally_df ['Purchase'] - final_test_tally_df ['Actual_Purchase']
    final_test_tally_df ['Error_sq'] = final_test_tally_df ['Error']*final_test_tally_df ['Error']
    rmse_test.append(np.sqrt(final_test_tally_df ['Error_sq'].sum()/final_test_tally_df.shape[0]))
    file_test = './Files_Folder/rmse/reg/Predicted_test_Result_SVD_mean_' + str(j+1) + '.csv'
    final_test_tally_df.to_csv(file_test, index=False)
    
reg = [0.01*(k+1) for k in range(20)]
colate = {'reg. param': reg, 'cv_mean': cv_mean, 'cv_std': cv_std, 'rmse_train': rmse_train, 'rmse_test':rmse_test}
colate_df = pd.DataFrame(colate)
colate_df.to_csv('./Files_Folder/rmse/reg/Result_With_Varying_reg_param_SVD.csv', index = False)
colate_df

new_prod = list(New_Prod_LUT.Product_ID)

cv_mean = []
cv_std = []
rmse_train = []
rmse_test = []

for j in range(10, 201, 10):
    
    algo_SVD = SVD(n_epochs=j, lr_all = 0.005, reg_all = 0.1)

    # A reader is still needed but only the rating_scale param is requiered.
    reader = Reader(rating_scale=(1, 5))

    # The columns must correspond to user id, item id and ratings (in that order).
    data = Dataset.load_from_df(User_Prod_NP[['User_ID', 'Product_ID', 'Normalized_Purchase']], reader)

    # We can now use this dataset as we please, e.g. calling cross_validate
    crva = cross_validate(algo_SVD, data, cv=3, verbose = True)

    cv_mean.append(crva['test_rmse'].mean())

    cv_std.append(crva['test_rmse'].std())
    
    # Using our model to predict the prices of the train data set
    # This is used to determine the training error
    train_user_list = []
    train_prod_list = []
    train_price = []
    for i in range(train_data.shape[0]):
        p1 = algo_SVD.predict(train_data.User_ID[i],train_data.Product_ID[i]).est
        rmin = train_Prod_sta.loc[3, train_data.Product_ID[i]]
        rmax = train_Prod_sta.loc[7, train_data.Product_ID[i]]
        dmin = 1
        dmax = 5
        p = (p1 - dmin)*(rmax - rmin)/(dmax - dmin) + rmin
        train_price.append(p)
        train_user_list.append(train_data.User_ID[i])
        train_prod_list.append(train_data.Product_ID[i])
    final_train_tally = {'User_ID':train_user_list, 'Product_ID':train_prod_list, 'Purchase': train_price}
    final_train_tally_df = pd.DataFrame(final_train_tally)
    
    
    test_user_list = []
    test_prod_list = []
    price = []
    for i in range(test_data_sub.shape[0]):
        if test_data_sub.Product_ID[i] in new_prod:
            p = New_Prod_LUT[New_Prod_LUT.Product_ID == test_data_sub.Product_ID[i]]['mean'].iloc[0]      
        else:
            p1 = algo_SVD.predict(test_data_sub.User_ID[i],test_data_sub.Product_ID[i]).est
            rmin = train_Prod_sta.loc[3, test_data_sub.Product_ID[i]]
            rmax = train_Prod_sta.loc[7, test_data_sub.Product_ID[i]]
            dmin = 1
            dmax = 5
            p = (p1 - dmin)*(rmax - rmin)/(dmax - dmin) + rmin
        price.append(p)
        test_user_list.append(test_data_sub.User_ID[i])
        test_prod_list.append(test_data_sub.Product_ID[i])
    final_test_tally = {'User_ID':test_user_list, 'Product_ID':test_prod_list, 'Purchase': price}
    final_test_tally_df = pd.DataFrame(final_test_tally)
    
    final_train_tally_df ['Actual_Purchase'] = train_data['Purchase']
    final_train_tally_df ['Error'] = final_train_tally_df ['Purchase'] - final_train_tally_df ['Actual_Purchase']
    final_train_tally_df ['Error_sq'] = final_train_tally_df ['Error']*final_train_tally_df ['Error']
    rmse_train.append(np.sqrt(final_train_tally_df ['Error_sq'].sum()/final_train_tally_df.shape[0]))    
    file_train = './Files_Folder/rmse/epoch/Predicted_train_Result_SVD_mean_' + str(j) + '.csv'
    final_train_tally_df.to_csv(file_train, index=False) 
    
    final_test_tally_df ['Actual_Purchase'] = test_data['Purchase']
    final_test_tally_df ['Error'] = final_test_tally_df ['Purchase'] - final_test_tally_df ['Actual_Purchase']
    final_test_tally_df ['Error_sq'] = final_test_tally_df ['Error']*final_test_tally_df ['Error']
    rmse_test.append(np.sqrt(final_test_tally_df ['Error_sq'].sum()/final_test_tally_df.shape[0]))
    file_test = './Files_Folder/rmse/epoch/Predicted_test_Result_SVD_mean_' + str(j) + '.csv'
    final_test_tally_df.to_csv(file_test, index=False)
    
epo = [k for k in range(10, 201, 10)]
colate = {'epoch': epo, 'cv_mean': cv_mean, 'cv_std': cv_std, 'rmse_train': rmse_train, 'rmse_test':rmse_test}
colate_df = pd.DataFrame(colate)
colate_df.to_csv('./Files_Folder/rmse/epoch/Result_With_Varying_reg_param_SVD.csv', index = False)
colate_df
